refactor: 검색 기능 훅으로 분리
