+++ frontend/src/hooks/queries/search/index.ts
@@ -1,1 +1,2 @@
 export { default as useSearchedProductsQuery } from './useSearchedProductsQuery';
+export { default as useSearch } from './useSearch';
